Implemented extended Storable Test/Validator and fixed several issues
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@45496 ebf7c1c2-ba36-0410-9fe8-c592906822b4

Former-commit-id: d268e798f625af047177961fc458d0d428832292

+++ src/org/jdownloader/api/captcha/CaptchaJob.java
@@ -19,7 +19,7 @@ public class CaptchaJob implements Storable {
 
     private String captchaCategory;
     private String explain;
-    private int    timeout   = -1;
+    private long   timeout   = -1;
     private long   created   = -1;
     private int    remaining = -1;
 
@@ -105,7 +105,7 @@ public class CaptchaJob implements Storable {
         return timeout;
     }
 
-    public void setTimeout(int timeout) {
+    public void setTimeout(long timeout) {
         this.timeout = timeout;
     }
